+++ app/Http/Controllers/CategoriesController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Category;
+use App\Models\Brand;
 use DataTables;
 
 class CategoriesController extends Controller
@@ -76,7 +77,8 @@ class CategoriesController extends Controller
      */
     public function show($id)
     {
-        //
+        $c = Category::find($id);
+        return response()->json($c, 200);
     }
 
     /**
@@ -93,13 +95,16 @@ class CategoriesController extends Controller
         Parámetros: recibe el parámetro 'id' de la Categoría a modificar y 'request' que tendrá los
         datos que se modificaron.
         Return: Devuelve un mensaje indicando que no existe la Categoría con ese 'id', o devuelve 
-        un mensaje indicando que si se pudo realizar la opreación de actualización */
+        un mensaje indicando que si se pudo realizar la opreación de actualización
+         */
 
         $category = Category::find($id);
         $validaData = $request->validate([
                             'category_name' => ['required','string']
                     ]);
-
+        
+        
+        
         if (!$category) {
             return response()->json([
                 'success' => false,
@@ -108,7 +113,7 @@ class CategoriesController extends Controller
         }
         
         $data = array(
-                'category_name' => $request->category_name,
+                'category_name' => strtoupper($request->category_name),
         );
 
         $updated = $category->update($data);
@@ -135,17 +140,49 @@ class CategoriesController extends Controller
         /*Esta función busca la Categoría con el id pasado por parámetro y lo elimina.
           Parámetros: recibe el parámetro id, que es el código de la categoría a eliminar  
           Return: devuelve un mensaje indicando si pudo eliminar la categoría o no */
-            
+          $product = Category::find($id)->products;
+
           $category = Category::find($id);
-            if ($category->delete()) {
-                return response()->json([
+
+          if (sizeof($product) > 0) {
+            return response()->json([
+                'success' => false,
+                'message' => 'La Categoría no se puede eliminar porque existe un producto asociado'
+            ], 500);
+          } 
+
+          if ($category->delete()) {
+            return response()->json([
                             'success' => true
                         ]);
-            } else {
+          } else {
                 return response()->json([
                             'success' => false,
                             'message' => 'La Categoría no pudo ser borrada'
             ], 500);
-                }
+            }
+    }
+
+    public function searchProducts($id){
+        $category = Category::find($id);
+        $products_category = Category::find($id)->products;
+        $products = array();
+        foreach ($products_category as $p) {
+            $brand = Brand::find($p->brand_id);
+            $data['product_id'] = $p->product_id;
+            $data['product_name'] = $p->product_name;
+            $data['product_description'] = $p->product_description;
+            $data['product_price'] = $p->product_price;
+            $data['product_image'] = $p->product_image;
+            $data['product_stock'] = $p->product_stock;
+            $data['product_offer_day'] = $p->product_offer_day;
+            $data['product_offer_day_order'] = $p->product_offer_day_order;
+            $data['product_best_seller'] = $p->product_best_seller;
+            $data['product_best_seller_order'] = $p->product_best_seller_order;
+            $data['category'] = $category;
+            $data['brand'] = $brand;
+            array_push($products,$data);
+        }
+        return response()->json($data, 200);
     }
 }
